+++ src/client/theme-default/index.ts
@@ -1,8 +1,15 @@
+import './styles/vars.css'
+import './styles/layout.css'
+import './styles/code.css'
+import './styles/custom-blocks.css'
+
 import Layout from './Layout.vue'
+import NotFound from './NotFound.vue'
 import { Theme } from '../app/theme'
 
 const theme: Theme = {
-  Layout
+  Layout,
+  NotFound
 }
 
 export default theme